[refactor] Drawable 인터페이스 삭제 및 Controller 리팩토링
Drawable 삭제
PlayerRule 삭제
DealerRule 삭제
User isDrawable을 abstract method로 두고, 각 자식 클래스에서 구현하도록 함
Player 기존 PlayerRule의 구현체를 옮겨 구현
Dealer 기존 DealerRule의 구현체를 옮겨 구현
Players 이름,상태 등을 목록으로 반환하는 메소드 추가
Controller Players를 뷰에 직접 전달하도록 로직 일부 수정
OutputView Players에서 목록을 받아 출력하도록 일부 메소드 수정
