--- src/main/java/domain/rule/Drawable.java
@@ -1,8 +1,0 @@
-package domain.rule;
-
-public interface Drawable {
-    public static final int ACE_BONUS_SCORE = 10;
-    public static final int BLACKJACK_FULL_SCORE = 21;
-
-    abstract boolean check(int score);
-}
