--- src/main/java/domain/rule/PlayerRule.java
@@ -1,7 +1,0 @@
-package domain.rule;
-
-public class PlayerRule implements Drawable{
-    public boolean check(int score) {
-        return score <= BLACKJACK_FULL_SCORE;
-    }
-}
