+++ src/main/java/domain/user/Dealer.java
@@ -1,8 +1,9 @@
 package domain.user;
 
-import static domain.rule.Drawable.BLACKJACK_FULL_SCORE;
+import static domain.user.HandCard.BLACKJACK_FULL_SCORE;
 
 public class Dealer extends User {
+    public static final int DRAW_MAX_SCORE = 16;
     private static final String DEALER_NAME = "딜러";
 
     public Dealer() {
@@ -18,5 +19,9 @@ public class Dealer extends User {
             return false;
         }
         return user.getScore() > BLACKJACK_FULL_SCORE || handCard.getScore() >= user.getScore();
+    }
+
+    public boolean isDrawable() {
+        return handCard.getScore() <= DRAW_MAX_SCORE;
     }
 }
